fix[client]: authentication schema fix
Initially I had intended to use one Zod validation schema throughout all of my authentication forms. However, upon creating the sign in form (already having the registration form created), the form would not submit, nor would it display anything in the console to direct me.
After some research, I discovered that the handleSubmit function given to us by the useForm hook wouldn't call the first argument (in this case, my onSubmit function) if the form failed validation. Having used the same validation for the registration form, I was surprised to find that this was the case.
In order to avoid this issue, I have changed the authentication schema into a registration schema and intend to create an individual schema for each form.
I will return to this later on, after completing the initial application, and find a way to patch this problem in a way that I can use one validation schema for all authentication forms (if Zod will allow me to do so).